Moved phone validations to their own file
